[REFACTOR][SCHEDULE] Replace Integer() trace boxing with IntImm in Schedule and MetaSchedule
Schedule trace attrs and MetaSchedule decision payloads boxed ints
into Integer() to pass through ffi::Any. IntImm(DataType::Int(32), N)
is the correct canonical form for integer constants in IR position;
Integer() was a redundant wrapper. Replace across traced_schedule.cc,
concrete_schedule, and ~15 MetaSchedule files.

Also change MultiLevelTilingWideVector's vector_length_in_bits parameter
from Integer to int64_t in both the header declaration and implementation,
matching the underlying field type.

+++ include/tvm/s_tir/meta_schedule/schedule_rule.h
@@ -231,7 +231,7 @@ class ScheduleRule : public ffi::ObjectRef {
    * \return The schedule rule created
    */
   TVM_DLL static ScheduleRule MultiLevelTilingWideVector(
-      ffi::String structure, Integer vector_length_in_bits,
+      ffi::String structure, int64_t vector_length_in_bits,
       ffi::Optional<int64_t> max_innermost_factor,
       ffi::Optional<ffi::Map<ffi::String, ffi::Any>> reuse_read,
       ffi::Optional<ffi::Map<ffi::String, ffi::Any>> reuse_write);
